Replaced size_t for unsigned int on ex01

+++ ex01/includes/iter.hpp
@@ -12,13 +12,12 @@
 
 
 #pragma once
-#include <cstddef> // unsigned int ( size_t )
 
 
 template <typename T>
-void iter( T* array, size_t length, void ( *f )( T& ) )
+void iter( T* array, unsigned int length, void ( *f )( T& ) )
 {
-  for( size_t i = 0; i < length; i++ ) {
+  for( unsigned int i = 0; i < length; i++ ) {
     f( array[i] );
   }
 }
